@@ -5,13 +5,14 @@ import 'package:notus/notus.dart';
 import 'package:quill_delta/quill_delta.dart';
 import 'package:test/test.dart';
 
+final rightAttrs = NotusStyle().merge(NotusAttribute.right);
 final ulAttrs = NotusStyle().merge(NotusAttribute.ul);
 final olAttrs = NotusStyle().merge(NotusAttribute.ol);
 final h1Attrs = NotusStyle().merge(NotusAttribute.h1);
 
 void main() {
   group('$BlockNode', () {
-    ContainerNode root;
+    late ContainerNode root;
     setUp(() {
       root = RootNode();
     });
@@ -39,8 +40,8 @@ void main() {
       root.retain(7, 1, ulAttrs);
       root.retain(13, 1, ulAttrs);
       expect(root.childCount, 1);
-      BlockNode block = root.first;
-      LineNode line = block.children.elementAt(1);
+      final block = root.first as BlockNode;
+      final line = block.children.elementAt(1) as LineNode;
       block.unwrapLine(line);
       expect(root.children, hasLength(3));
       expect(root.children.elementAt(0), const TypeMatcher<BlockNode>());
@@ -53,12 +54,12 @@ void main() {
       root.retain(11, 1, ulAttrs);
 
       expect(root.childCount, 1);
-      BlockNode block = root.first;
+      final block = root.first as BlockNode;
       expect(block.style.get(NotusAttribute.block), NotusAttribute.ul);
       expect(block.childCount, 1);
       expect(block.first, const TypeMatcher<LineNode>());
 
-      LineNode line = block.first;
+      final line = block.first as LineNode;
       final delta = Delta()
         ..insert('Hello world')
         ..insert('\n', ulAttrs.toJson());
@@ -70,10 +71,24 @@ void main() {
       root.retain(21, 1, ulAttrs);
 
       expect(root.childCount, 2);
-      BlockNode block = root.last;
+      final block = root.last as BlockNode;
       expect(block.style.get(NotusAttribute.block), NotusAttribute.ul);
       expect(block.childCount, 1);
       expect(block.first, const TypeMatcher<LineNode>());
+    });
+
+    test('format first line as list and right aligned', () {
+      root.insert(0, 'Hello world\nAb cd ef!', null);
+      root.retain(11, 1, rightAttrs);
+      root.retain(11, 1, ulAttrs);
+
+      expect(root.childCount, 2);
+      final block = root.first as BlockNode;
+      expect(block.style.get(NotusAttribute.block), NotusAttribute.ul);
+      expect(block.childCount, 1);
+      expect(block.first, const TypeMatcher<LineNode>());
+      final line = block.first as LineNode;
+      expect(line.style.get(NotusAttribute.alignment), NotusAttribute.right);
     });
 
     test('format two sibling lines as list', () {
@@ -82,7 +97,7 @@ void main() {
       root.retain(21, 1, ulAttrs);
 
       expect(root.childCount, 1);
-      BlockNode block = root.first;
+      final block = root.first as BlockNode;
       expect(block.style.get(NotusAttribute.block), NotusAttribute.ul);
       expect(block.childCount, 2);
       expect(block.first, const TypeMatcher<LineNode>());
@@ -155,7 +170,7 @@ void main() {
       expect(root.toDelta(), expected);
     });
 
-    test('insert line-break at the begining of the document', () {
+    test('insert line-break at the beginning of the document', () {
       root.insert(
           0, 'London Grammar Songs\nHey now\nStrong\nIf You Wait', null);
       root.retain(20, 1, ulAttrs);
