@@ -1,45 +1,164 @@
 import 'package:flutter/material.dart';
+import 'package:notus/notus.dart';
 
 import 'controller.dart';
 import 'editor.dart';
-import 'image.dart';
-import 'mode.dart';
-import 'toolbar.dart';
 
-/// Zefyr editor with material design decorations.
 class ZefyrField extends StatefulWidget {
-  /// Decoration to paint around this editor.
-  final InputDecoration decoration;
-
-  /// Height of this editor field.
-  final double height;
+  /// Controller object which establishes a link between a rich text document
+  /// and this editor.
+  ///
+  /// Must not be null.
   final ZefyrController controller;
-  final FocusNode focusNode;
+
+  /// Controls whether this editor has keyboard focus.
+  ///
+  /// Can be `null` in which case this editor creates its own instance to
+  /// control keyboard focus.
+  final FocusNode? focusNode;
+
+  /// The [ScrollController] to use when vertically scrolling the contents.
+  ///
+  /// If `null` then this editor instantiates a new ScrollController.
+  ///
+  /// Scroll controller must not be `null` if [scrollable] is set to `false`.
+  final ScrollController? scrollController;
+
+  /// Whether this editor should create a scrollable container for its content.
+  ///
+  /// When set to `true` the editor's height can be controlled by [minHeight],
+  /// [maxHeight] and [expands] properties.
+  ///
+  /// When set to `false` the editor always expands to fit the entire content
+  /// of the document and should normally be placed as a child of another
+  /// scrollable widget, otherwise the content may be clipped.
+  ///
+  /// The [scrollController] property must not be `null` when this is set to
+  /// `false`.
+  ///
+  /// Set to `true` by default.
+  final bool scrollable;
+
+  /// Additional space around the content of this editor.
+  final EdgeInsetsGeometry padding;
+
+  /// Whether this editor should focus itself if nothing else is already
+  /// focused.
+  ///
+  /// If true, the keyboard will open as soon as this editor obtains focus.
+  /// Otherwise, the keyboard is only shown after the user taps the editor.
+  ///
+  /// Defaults to `false`. Cannot be `null`.
   final bool autofocus;
-  final ZefyrMode mode;
-  final ZefyrToolbarDelegate toolbarDelegate;
-  final ZefyrImageDelegate imageDelegate;
-  final ScrollPhysics physics;
+
+  /// Whether to show cursor.
+  ///
+  /// The cursor refers to the blinking caret when the editor is focused.
+  final bool showCursor;
+
+  /// Whether the text can be changed.
+  ///
+  /// When this is set to `true`, the text cannot be modified
+  /// by any shortcut or keyboard operation. The text is still selectable.
+  ///
+  /// Defaults to `false`. Must not be `null`.
+  final bool readOnly;
+
+  /// Whether to enable user interface affordances for changing the
+  /// text selection.
+  ///
+  /// For example, setting this to true will enable features such as
+  /// long-pressing the editor to select text and show the
+  /// cut/copy/paste menu, and tapping to move the text cursor.
+  ///
+  /// When this is false, the text selection cannot be adjusted by
+  /// the user, text cannot be copied, and the user cannot paste into
+  /// the text field from the clipboard.
+  final bool enableInteractiveSelection;
+
+  /// The minimum height to be occupied by this editor.
+  ///
+  /// This only has effect if [scrollable] is set to `true` and [expands] is
+  /// set to `false`.
+  final double? minHeight;
+
+  /// The maximum height to be occupied by this editor.
+  ///
+  /// This only has effect if [scrollable] is set to `true` and [expands] is
+  /// set to `false`.
+  final double? maxHeight;
+
+  /// Whether this editor's height will be sized to fill its parent.
+  ///
+  /// This only has effect if [scrollable] is set to `true`.
+  ///
+  /// If expands is set to true and wrapped in a parent widget like [Expanded]
+  /// or [SizedBox], the editor will expand to fill the parent.
+  ///
+  /// [maxHeight] and [minHeight] must both be `null` when this is set to
+  /// `true`.
+  ///
+  /// Defaults to `false`.
+  final bool expands;
+
+  /// Configures how the platform keyboard will select an uppercase or
+  /// lowercase keyboard.
+  ///
+  /// Only supports text keyboards, other keyboard types will ignore this
+  /// configuration. Capitalization is locale-aware.
+  ///
+  /// Defaults to [TextCapitalization.sentences]. Must not be `null`.
+  final TextCapitalization textCapitalization;
 
   /// The appearance of the keyboard.
   ///
   /// This setting is only honored on iOS devices.
   ///
-  /// If unset, defaults to the brightness of [ThemeData.primaryColorBrightness].
+  /// Defaults to [Brightness.light].
   final Brightness keyboardAppearance;
 
+  /// The [ScrollPhysics] to use when vertically scrolling the input.
+  ///
+  /// This only has effect if [scrollable] is set to `true`.
+  ///
+  /// If not specified, it will behave according to the current platform.
+  ///
+  /// See [Scrollable.physics].
+  final ScrollPhysics? scrollPhysics;
+
+  /// Callback to invoke when user wants to launch a URL.
+  final ValueChanged<String?>? onLaunchUrl;
+
+  final InputDecoration? decoration;
+
+  final Widget? toolbar;
+
+  /// Builder function for embeddable objects.
+  ///
+  /// Defaults to [defaultZefyrEmbedBuilder].
+  final ZefyrEmbedBuilder embedBuilder;
+
   const ZefyrField({
-    Key key,
-    this.decoration,
-    this.height,
-    this.controller,
+    Key? key,
+    required this.controller,
     this.focusNode,
+    this.scrollController,
+    this.scrollable = true,
+    this.padding = EdgeInsets.zero,
     this.autofocus = false,
-    this.mode,
-    this.toolbarDelegate,
-    this.imageDelegate,
-    this.physics,
-    this.keyboardAppearance,
+    this.showCursor = true,
+    this.readOnly = false,
+    this.enableInteractiveSelection = true,
+    this.minHeight,
+    this.maxHeight,
+    this.expands = false,
+    this.textCapitalization = TextCapitalization.sentences,
+    this.keyboardAppearance = Brightness.light,
+    this.scrollPhysics,
+    this.onLaunchUrl,
+    this.decoration,
+    this.toolbar,
+    this.embedBuilder = defaultZefyrEmbedBuilder,
   }) : super(key: key);
 
   @override
@@ -47,36 +166,76 @@ class ZefyrField extends StatefulWidget {
 }
 
 class _ZefyrFieldState extends State<ZefyrField> {
-  ZefyrMode get _effectiveMode => widget.mode ?? ZefyrMode.edit;
+  late bool _focused;
+
+  void _editorFocusChanged() {
+    setState(() {
+      _focused = widget.focusNode!.hasFocus;
+    });
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _focused = widget.focusNode!.hasFocus;
+    widget.focusNode!.addListener(_editorFocusChanged);
+  }
+
+  @override
+  void didUpdateWidget(covariant ZefyrField oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.focusNode != oldWidget.focusNode) {
+      oldWidget.focusNode!.removeListener(_editorFocusChanged);
+      widget.focusNode!.addListener(_editorFocusChanged);
+      _focused = widget.focusNode!.hasFocus;
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     Widget child = ZefyrEditor(
-      padding: EdgeInsets.symmetric(vertical: 6.0),
       controller: widget.controller,
       focusNode: widget.focusNode,
+      scrollController: widget.scrollController,
+      scrollable: widget.scrollable,
+      padding: widget.padding,
       autofocus: widget.autofocus,
-      mode: _effectiveMode,
-      toolbarDelegate: widget.toolbarDelegate,
-      imageDelegate: widget.imageDelegate,
-      physics: widget.physics,
+      showCursor: widget.showCursor,
+      readOnly: widget.readOnly,
+      enableInteractiveSelection: widget.enableInteractiveSelection,
+      minHeight: widget.minHeight,
+      maxHeight: widget.maxHeight,
+      expands: widget.expands,
+      textCapitalization: widget.textCapitalization,
       keyboardAppearance: widget.keyboardAppearance,
+      scrollPhysics: widget.scrollPhysics,
+      onLaunchUrl: widget.onLaunchUrl,
+      embedBuilder: widget.embedBuilder,
     );
 
-    if (widget.height != null) {
-      child = ConstrainedBox(
-        constraints: BoxConstraints.tightFor(height: widget.height),
-        child: child,
+    if (widget.toolbar != null) {
+      child = Column(
+        children: [
+          child,
+          Visibility(
+            visible: _focused,
+            maintainSize: true,
+            maintainAnimation: true,
+            maintainState: true,
+            child: widget.toolbar!,
+          ),
+        ],
       );
     }
 
     return AnimatedBuilder(
       animation:
-          Listenable.merge(<Listenable>[widget.focusNode, widget.controller]),
-      builder: (BuildContext context, Widget child) {
+          Listenable.merge(<Listenable?>[widget.focusNode, widget.controller]),
+      builder: (BuildContext context, Widget? child) {
         return InputDecorator(
           decoration: _getEffectiveDecoration(),
-          isFocused: widget.focusNode.hasFocus,
-          isEmpty: widget.controller.document.length == 1,
+          isFocused: widget.focusNode!.hasFocus,
+          isEmpty: _isEmpty,
           child: child,
         );
       },
@@ -84,11 +243,23 @@ class _ZefyrFieldState extends State<ZefyrField> {
     );
   }
 
+  /// Field is considered empty when the document consists of
+  /// a single empty line of text with no styles applied to it
+  bool get _isEmpty {
+    if (widget.controller.document.length > 1) {
+      return false;
+    }
+    final node = widget.controller.document.root.first;
+    assert(node is StyledNode);
+    return (node as StyledNode).style.isEmpty;
+  }
+
   InputDecoration _getEffectiveDecoration() {
     final effectiveDecoration = (widget.decoration ?? const InputDecoration())
         .applyDefaults(Theme.of(context).inputDecorationTheme)
         .copyWith(
-          enabled: _effectiveMode == ZefyrMode.edit,
+          enabled: !widget.readOnly,
+          hintMaxLines: widget.decoration?.hintMaxLines,
         );
 
     return effectiveDecoration;
