@@ -1,24 +1,31 @@
+// Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 import 'package:flutter/material.dart';
+import 'package:zefyr/zefyr.dart';
 
 /// Applies a Zefyr editor theme to descendant widgets.
 ///
-/// Describes colors and typographic styles.
+/// Describes colors and typographic styles for an editor.
 ///
 /// Descendant widgets obtain the current theme's [ZefyrThemeData] object using
 /// [ZefyrTheme.of].
+///
+/// See also:
+///
+///   * [ZefyrThemeData], which describes actual configuration of a theme.
 class ZefyrTheme extends InheritedWidget {
   final ZefyrThemeData data;
 
   /// Applies the given theme [data] to [child].
   ///
   /// The [data] and [child] arguments must not be null.
-  ZefyrTheme({
-    Key key,
-    @required this.data,
-    @required Widget child,
-  })  : assert(data != null),
-        assert(child != null),
-        super(key: key, child: child);
+  const ZefyrTheme({
+    Key? key,
+    required this.data,
+    required Widget child,
+  }) : super(key: key, child: child);
 
   @override
   bool updateShouldNotify(ZefyrTheme oldWidget) {
@@ -31,545 +38,330 @@ class ZefyrTheme extends InheritedWidget {
   /// Returns `null` if there is no [ZefyrTheme] in the given build context
   /// and [nullOk] is set to `true`. If [nullOk] is set to `false` (default)
   /// then this method asserts.
-  static ZefyrThemeData of(BuildContext context, {bool nullOk = false}) {
+  static ZefyrThemeData? of(BuildContext context, {bool nullOk = false}) {
     final widget = context.dependOnInheritedWidgetOfExactType<ZefyrTheme>();
     if (widget == null && nullOk) return null;
     assert(widget != null,
         '$ZefyrTheme.of() called with a context that does not contain a ZefyrEditor.');
-    return widget.data;
+    return widget!.data;
   }
 }
 
-/// Holds colors and typography values for a Zefyr design theme.
-///
-/// To obtain the current theme, use [ZefyrTheme.of].
-@immutable
+/// Vertical spacing around a block of text.
+class VerticalSpacing {
+  final double top;
+  final double bottom;
+
+  const VerticalSpacing({this.top = 0.0, this.bottom = 0.0});
+
+  const VerticalSpacing.zero()
+      : top = 0.0,
+        bottom = 0.0;
+}
+
 class ZefyrThemeData {
-  /// Default theme used for document lines in Zefyr editor.
-  ///
-  /// Defines text style and spacing for regular paragraphs of text with
-  /// no style attributes applied.
-  final LineTheme defaultLineTheme;
-
-  /// The text styles, padding and decorations used to render text with
-  /// different style attributes.
-  final AttributeTheme attributeTheme;
-
-  /// The width of indentation used for blocks (lists, quotes, code).
-  final double indentWidth;
-
-  /// The colors used to render editor toolbar.
-  final ToolbarTheme toolbarTheme;
-
-  /// Creates a [ZefyrThemeData] given a set of exact values.
-  const ZefyrThemeData({
-    this.defaultLineTheme,
-    this.attributeTheme,
-    this.indentWidth,
-    this.toolbarTheme,
-  });
-
-  /// The default editor theme.
-  factory ZefyrThemeData.fallback(BuildContext context) {
-    final defaultStyle = DefaultTextStyle.of(context);
-    final defaultLineTheme = LineTheme(
-      textStyle: defaultStyle.style.copyWith(
-        fontSize: 16.0,
-        height: 1.3,
-      ),
-      padding: EdgeInsets.symmetric(vertical: 8.0),
-    );
-    return ZefyrThemeData(
-      defaultLineTheme: defaultLineTheme,
-      attributeTheme: AttributeTheme.fallback(context, defaultLineTheme),
-      indentWidth: 16.0,
-      toolbarTheme: ToolbarTheme.fallback(context),
-    );
-  }
-
-  /// Creates a copy of this theme but with the given fields replaced with
-  /// the new values.
-  ZefyrThemeData copyWith({
-    LineTheme defaultLineTheme,
-    AttributeTheme attributeTheme,
-    double indentWidth,
-    ToolbarTheme toolbarTheme,
-  }) {
-    return ZefyrThemeData(
-      defaultLineTheme: defaultLineTheme ?? this.defaultLineTheme,
-      attributeTheme: attributeTheme ?? this.attributeTheme,
-      indentWidth: indentWidth ?? this.indentWidth,
-      toolbarTheme: toolbarTheme ?? this.toolbarTheme,
-    );
-  }
-
-  /// Creates a new [ZefyrThemeData] where each property from this object has
-  /// been merged with the matching text style from the `other` object.
-  ZefyrThemeData merge(ZefyrThemeData other) {
-    if (other == null) return this;
-    return copyWith(
-      defaultLineTheme: defaultLineTheme?.merge(other.defaultLineTheme) ??
-          other.defaultLineTheme,
-      attributeTheme:
-          attributeTheme?.merge(other.attributeTheme) ?? other.attributeTheme,
-      indentWidth: other.indentWidth ?? indentWidth,
-      toolbarTheme:
-          toolbarTheme?.merge(other.toolbarTheme) ?? other.toolbarTheme,
-    );
-  }
-
-  @override
-  bool operator ==(other) {
-    if (other.runtimeType != runtimeType) return false;
-    final ZefyrThemeData otherData = other;
-    return (otherData.defaultLineTheme == defaultLineTheme) &&
-        (otherData.attributeTheme == attributeTheme) &&
-        (otherData.indentWidth == indentWidth) &&
-        (otherData.toolbarTheme == toolbarTheme);
-  }
-
-  @override
-  int get hashCode {
-    return hashList([
-      defaultLineTheme,
-      attributeTheme,
-      indentWidth,
-      toolbarTheme,
-    ]);
-  }
-}
-
-/// Holds typography values for a document line in Zefyr editor.
-///
-/// Applicable for regular paragraphs, headings and lines within blocks
-/// (lists, quotes). Blocks may override some of these values using [BlockTheme].
-@immutable
-class LineTheme {
-  /// Default text style for a document line.
-  final TextStyle textStyle;
-
-  /// Additional space around a document line.
-  final EdgeInsets padding;
-
-  /// Creates a [LineTheme] given a set of exact values.
-  LineTheme({@required this.textStyle, @required this.padding})
-      : assert(textStyle != null),
-        assert(padding != null);
-
-  /// Creates a copy of this theme but with the given fields replaced with
-  /// the new values.
-  LineTheme copyWith({TextStyle textStyle, EdgeInsets padding}) {
-    return LineTheme(
-      textStyle: textStyle ?? this.textStyle,
-      padding: padding ?? this.padding,
-    );
-  }
-
-  /// Creates a new [LineTheme] where each property from this object has
-  /// been merged with the matching property from the `other` object.
-  ///
-  /// Text style is merged using [TextStyle.merge] when this and other
-  /// theme have this value set.
-  ///
-  /// If padding property is set in other then it replaces value of this
-  /// theme.
-  LineTheme merge(LineTheme other) {
-    if (other == null) return this;
-    return copyWith(
-      textStyle: textStyle?.merge(other.textStyle) ?? other.textStyle,
-      padding: other.padding ?? padding,
-    );
-  }
-
-  @override
-  bool operator ==(other) {
-    if (other.runtimeType != runtimeType) return false;
-    final LineTheme otherTheme = other;
-    return (otherTheme.textStyle == textStyle) &&
-        (otherTheme.padding == padding);
-  }
-
-  @override
-  int get hashCode => hashValues(textStyle, padding);
-}
-
-/// Holds typography values for a block of lines in Zefyr editor.
-@immutable
-class BlockTheme {
-  /// Default text style for all text within a block, can be null.
-  ///
-  /// Takes precedence over line-level text style set by [LineTheme] if
-  /// [inheritLineTextStyle] is set to `false`. Otherwise this text style
-  /// is merged with the line's style.
-  final TextStyle textStyle;
-
-  /// Whether [textStyle] specified by this block theme should be merged with
-  /// text style of each individual line.
-  ///
-  /// Only applicable if [textStyle] is not null.
-  ///
-  /// If set to `true` then [textStyle] is merged with text style specified
-  /// by [LineTheme] of each line within a block. Otherwise [textStyle]
-  /// takes precedence and replaces style of [LineTheme].
-  final bool inheritLineTextStyle;
-
-  /// Space around the block.
-  final EdgeInsets padding;
-
-  /// Space around each individual line within a block, can be null.
-  ///
-  /// Takes precedence over line padding set in [LineTheme].
-  final EdgeInsets linePadding;
-
-  /// Creates a [BlockTheme] given a set of exact values.
-  const BlockTheme({
-    this.textStyle,
-    this.inheritLineTextStyle = true,
-    this.padding,
-    this.linePadding,
-  });
-
-  /// Creates a copy of this theme but with the given fields replaced with
-  /// the new values.
-  BlockTheme copyWith({
-    TextStyle textStyle,
-    EdgeInsets padding,
-    bool inheritLineTextStyle,
-    EdgeInsets linePadding,
-  }) {
-    return BlockTheme(
-      textStyle: textStyle ?? this.textStyle,
-      inheritLineTextStyle: inheritLineTextStyle ?? this.inheritLineTextStyle,
-      padding: padding ?? this.padding,
-      linePadding: linePadding ?? this.linePadding,
-    );
-  }
-
-  /// Creates a new [BlockTheme] where each property from this object has
-  /// been merged with the matching property from the `other` object.
-  ///
-  /// Text style is merged using [TextStyle.merge] when this and other
-  /// theme have this field set.
-  ///
-  /// If padding property is set in other then it replaces value of this
-  /// theme. [linePadding] follows the same logic.
-  BlockTheme merge(BlockTheme other) {
-    if (other == null) return this;
-    return copyWith(
-      textStyle: textStyle?.merge(other.textStyle) ?? other.textStyle,
-      inheritLineTextStyle: other.inheritLineTextStyle ?? inheritLineTextStyle,
-      padding: other.padding ?? padding,
-      linePadding: other.linePadding ?? linePadding,
-    );
-  }
-
-  @override
-  bool operator ==(other) {
-    if (other.runtimeType != runtimeType) return false;
-    final BlockTheme otherTheme = other;
-    return (otherTheme.textStyle == textStyle) &&
-        (otherTheme.inheritLineTextStyle == inheritLineTextStyle) &&
-        (otherTheme.padding == padding) &&
-        (otherTheme.linePadding == linePadding);
-  }
-
-  @override
-  int get hashCode =>
-      hashValues(textStyle, inheritLineTextStyle, padding, linePadding);
-}
-
-/// Holds style information for all format attributes supported by Zefyr editor.
-@immutable
-class AttributeTheme {
-  /// Style used to render "bold" text.
+  /// Style of bold text.
   final TextStyle bold;
 
-  /// Style used to render "italic" text.
+  /// Style of italic text.
   final TextStyle italic;
 
-  /// Style used to render "underline" text.
+  /// Style of underline text.
   final TextStyle underline;
 
-  /// Style used to render text containing links.
+  /// Style of strikethrough text.
+  final TextStyle strikethrough;
+
+  /// Theme of inline code.
+  final InlineCodeThemeData inlineCode;
+
+  /// Style of links in text.
   final TextStyle link;
 
-  /// Style theme used to render largest headings.
-  final LineTheme heading1;
+  /// Default style theme for regular paragraphs of text.
+  final TextBlockTheme paragraph; // spacing: top: 6, bottom: 10
+  /// Style theme for level 1 headings.
+  final TextBlockTheme heading1;
 
-  /// Style theme used to render medium headings.
-  final LineTheme heading2;
+  /// Style theme for level 2 headings.
+  final TextBlockTheme heading2;
 
-  /// Style theme used to render smaller headings.
-  final LineTheme heading3;
+  /// Style theme for level 3 headings.
+  final TextBlockTheme heading3;
 
-  /// Style theme used to render bullet lists.
-  final BlockTheme bulletList;
+  /// Style theme for bullet and number lists.
+  final TextBlockTheme lists;
 
-  /// Style theme used to render number lists.
-  final BlockTheme numberList;
+  /// Style theme for quote blocks.
+  final TextBlockTheme quote;
 
-  /// Style theme used to render quote blocks.
-  final BlockTheme quote;
+  /// Style theme for code blocks.
+  final TextBlockTheme code;
 
-  /// Style theme used to render code blocks.
-  final BlockTheme code;
-
-  /// Creates a [AttributeTheme] given a set of exact values.
-  AttributeTheme({
-    this.bold,
-    this.italic,
-    this.underline,
-    this.link,
-    this.heading1,
-    this.heading2,
-    this.heading3,
-    this.bulletList,
-    this.numberList,
-    this.quote,
-    this.code,
+  ZefyrThemeData({
+    required this.bold,
+    required this.italic,
+    required this.underline,
+    required this.strikethrough,
+    required this.inlineCode,
+    required this.link,
+    required this.paragraph,
+    required this.heading1,
+    required this.heading2,
+    required this.heading3,
+    required this.lists,
+    required this.quote,
+    required this.code,
   });
 
-  /// The default attribute theme.
-  factory AttributeTheme.fallback(
-      BuildContext context, LineTheme defaultLineTheme) {
-    final theme = Theme.of(context);
+  factory ZefyrThemeData.fallback(BuildContext context) {
+    final themeData = Theme.of(context);
+    final defaultStyle = DefaultTextStyle.of(context);
+    final baseStyle = defaultStyle.style.copyWith(
+      fontSize: 16.0,
+      height: 1.3,
+    );
+    const baseSpacing = VerticalSpacing(top: 6.0, bottom: 10);
 
-    String monospaceFontFamily;
-    switch (theme.platform) {
+    String fontFamily;
+    switch (themeData.platform) {
       case TargetPlatform.iOS:
       case TargetPlatform.macOS:
-        monospaceFontFamily = 'Menlo';
+        fontFamily = 'Menlo';
         break;
       case TargetPlatform.android:
       case TargetPlatform.fuchsia:
-        monospaceFontFamily = 'Roboto Mono';
+      case TargetPlatform.windows:
+      case TargetPlatform.linux:
+        fontFamily = 'Roboto Mono';
         break;
-      default:
-        throw UnimplementedError('Platform ${theme.platform} not implemented.');
     }
 
-    return AttributeTheme(
-      bold: TextStyle(fontWeight: FontWeight.bold),
-      italic: TextStyle(fontStyle: FontStyle.italic),
-      underline: TextStyle(decoration: TextDecoration.underline),
-      link: TextStyle(
-        decoration: TextDecoration.underline,
-        color: theme.accentColor,
+    final inlineCodeStyle = TextStyle(
+      fontSize: 14,
+      color: themeData.colorScheme.primaryVariant.withOpacity(0.8),
+      fontFamily: fontFamily,
+    );
+
+    return ZefyrThemeData(
+      bold: const TextStyle(fontWeight: FontWeight.bold),
+      italic: const TextStyle(fontStyle: FontStyle.italic),
+      underline: const TextStyle(decoration: TextDecoration.underline),
+      strikethrough: const TextStyle(decoration: TextDecoration.lineThrough),
+      inlineCode: InlineCodeThemeData(
+        backgroundColor: Colors.grey.shade100,
+        radius: const Radius.circular(3),
+        style: inlineCodeStyle,
+        heading1: inlineCodeStyle.copyWith(
+          fontSize: 32,
+          fontWeight: FontWeight.w300,
+        ),
+        heading2: inlineCodeStyle.copyWith(fontSize: 22),
+        heading3: inlineCodeStyle.copyWith(
+          fontSize: 18,
+          fontWeight: FontWeight.w500,
+        ),
       ),
-      heading1: LineTheme(
-        textStyle: defaultLineTheme.textStyle.copyWith(
+      link: TextStyle(
+        color: themeData.colorScheme.primaryVariant,
+        decoration: TextDecoration.underline,
+      ),
+      paragraph: TextBlockTheme(
+        style: baseStyle,
+        spacing: baseSpacing,
+        // lineSpacing is not relevant for paragraphs since they consist of one line
+      ),
+      heading1: TextBlockTheme(
+        style: defaultStyle.style.copyWith(
           fontSize: 34.0,
-          color: defaultLineTheme.textStyle.color.withOpacity(0.7),
+          color: defaultStyle.style.color?.withOpacity(0.70),
           height: 1.15,
           fontWeight: FontWeight.w300,
         ),
-        padding: EdgeInsets.only(top: 16.0),
+        spacing: const VerticalSpacing(top: 16.0, bottom: 0.0),
       ),
-      heading2: LineTheme(
-        textStyle: defaultLineTheme.textStyle.copyWith(
+      heading2: TextBlockTheme(
+        style: TextStyle(
           fontSize: 24.0,
-          color: defaultLineTheme.textStyle.color.withOpacity(0.7),
+          color: defaultStyle.style.color?.withOpacity(0.70),
           height: 1.15,
           fontWeight: FontWeight.normal,
         ),
-        padding: EdgeInsets.only(top: 8.0),
+        spacing: const VerticalSpacing(bottom: 0.0, top: 8.0),
       ),
-      heading3: LineTheme(
-        textStyle: defaultLineTheme.textStyle.copyWith(
+      heading3: TextBlockTheme(
+        style: TextStyle(
           fontSize: 20.0,
-          color: defaultLineTheme.textStyle.color.withOpacity(0.7),
-          height: 1.15,
+          color: defaultStyle.style.color?.withOpacity(0.70),
+          height: 1.25,
           fontWeight: FontWeight.w500,
         ),
-        padding: EdgeInsets.only(top: 8.0),
+        spacing: const VerticalSpacing(bottom: 0.0, top: 8.0),
       ),
-      bulletList: BlockTheme(
-        padding: EdgeInsets.symmetric(vertical: 8.0),
-        linePadding: EdgeInsets.symmetric(vertical: 2.0),
+      lists: TextBlockTheme(
+        style: baseStyle,
+        spacing: baseSpacing,
+        lineSpacing: const VerticalSpacing(bottom: 0),
       ),
-      numberList: BlockTheme(
-        padding: EdgeInsets.symmetric(vertical: 8.0),
-        linePadding: EdgeInsets.symmetric(vertical: 2.0),
-      ),
-      quote: BlockTheme(
-        padding: EdgeInsets.symmetric(vertical: 8.0),
-        textStyle: TextStyle(
-          color: defaultLineTheme.textStyle.color.withOpacity(0.6),
+      quote: TextBlockTheme(
+        style: TextStyle(color: baseStyle.color?.withOpacity(0.6)),
+        spacing: baseSpacing,
+        lineSpacing: const VerticalSpacing(top: 6, bottom: 2),
+        decoration: BoxDecoration(
+          border: BorderDirectional(
+            start: BorderSide(width: 4, color: Colors.grey.shade300),
+          ),
         ),
-        inheritLineTextStyle: true,
       ),
-      code: BlockTheme(
-        padding: EdgeInsets.symmetric(vertical: 8.0),
-        textStyle: TextStyle(
-          fontFamily: monospaceFontFamily,
-          fontSize: 14.0,
-          color: defaultLineTheme.textStyle.color.withOpacity(0.8),
-          height: 1.25,
+      code: TextBlockTheme(
+        style: TextStyle(
+          color: Colors.blue.shade900.withOpacity(0.9),
+          fontFamily: fontFamily,
+          fontSize: 13.0,
+          height: 1.4,
         ),
-        inheritLineTextStyle: false,
-        linePadding: EdgeInsets.zero,
+        spacing: baseSpacing,
+        decoration: BoxDecoration(
+          color: Colors.grey.shade50,
+          borderRadius: BorderRadius.circular(2),
+        ),
       ),
     );
   }
 
-  /// Creates a new [AttributeTheme] where each property from this object has
-  /// been merged with the matching property from the `other` object.
-  AttributeTheme copyWith({
-    TextStyle bold,
-    TextStyle italic,
-    TextStyle underline,
-    TextStyle link,
-    LineTheme heading1,
-    LineTheme heading2,
-    LineTheme heading3,
-    BlockTheme bulletList,
-    BlockTheme numberList,
-    BlockTheme quote,
-    BlockTheme code,
+  ZefyrThemeData copyWith({
+    TextStyle? bold,
+    TextStyle? italic,
+    TextStyle? underline,
+    TextStyle? strikethrough,
+    TextStyle? link,
+    InlineCodeThemeData? inlineCode,
+    TextBlockTheme? paragraph,
+    TextBlockTheme? heading1,
+    TextBlockTheme? heading2,
+    TextBlockTheme? heading3,
+    TextBlockTheme? lists,
+    TextBlockTheme? quote,
+    TextBlockTheme? code,
   }) {
-    return AttributeTheme(
+    return ZefyrThemeData(
       bold: bold ?? this.bold,
       italic: italic ?? this.italic,
       underline: underline ?? this.underline,
+      strikethrough: strikethrough ?? this.strikethrough,
+      inlineCode: inlineCode ?? this.inlineCode,
       link: link ?? this.link,
+      paragraph: paragraph ?? this.paragraph,
       heading1: heading1 ?? this.heading1,
       heading2: heading2 ?? this.heading2,
       heading3: heading3 ?? this.heading3,
-      bulletList: bulletList ?? this.bulletList,
-      numberList: numberList ?? this.numberList,
+      lists: lists ?? this.lists,
       quote: quote ?? this.quote,
       code: code ?? this.code,
     );
   }
 
-  /// Creates a new [AttributeTheme] where each property from this object has
-  /// been merged with the matching property from the `other` object.
-  AttributeTheme merge(AttributeTheme other) {
-    if (other == null) return this;
+  ZefyrThemeData merge(ZefyrThemeData other) {
     return copyWith(
-      bold: bold?.merge(other.bold) ?? other.bold,
-      italic: italic?.merge(other.italic) ?? other.italic,
-      underline: underline?.merge(other.underline) ?? other.underline,
-      link: link?.merge(other.link) ?? other.link,
-      heading1: heading1?.merge(other.heading1) ?? other.heading1,
-      heading2: heading2?.merge(other.heading2) ?? other.heading2,
-      heading3: heading3?.merge(other.heading3) ?? other.heading3,
-      bulletList: bulletList?.merge(other.bulletList) ?? other.bulletList,
-      numberList: numberList?.merge(other.numberList) ?? other.numberList,
-      quote: quote?.merge(other.quote) ?? other.quote,
-      code: code?.merge(other.code) ?? other.code,
+      bold: other.bold,
+      italic: other.italic,
+      underline: other.underline,
+      strikethrough: other.strikethrough,
+      inlineCode: other.inlineCode,
+      link: other.link,
+      paragraph: other.paragraph,
+      heading1: other.heading1,
+      heading2: other.heading2,
+      heading3: other.heading3,
+      lists: other.lists,
+      quote: other.quote,
+      code: other.code,
     );
-  }
-
-  @override
-  bool operator ==(other) {
-    if (other.runtimeType != runtimeType) return false;
-    final AttributeTheme otherTheme = other;
-    return (otherTheme.bold == bold) &&
-        (otherTheme.italic == italic) &&
-        (otherTheme.underline == underline) &&
-        (otherTheme.link == link) &&
-        (otherTheme.heading1 == heading1) &&
-        (otherTheme.heading2 == heading2) &&
-        (otherTheme.heading3 == heading3) &&
-        (otherTheme.bulletList == bulletList) &&
-        (otherTheme.numberList == numberList) &&
-        (otherTheme.quote == quote) &&
-        (otherTheme.code == code);
-  }
-
-  @override
-  int get hashCode {
-    return hashList([
-      bold,
-      italic,
-      link,
-      heading1,
-      heading2,
-      heading3,
-      bulletList,
-      numberList,
-      quote,
-      code,
-    ]);
   }
 }
 
-/// Defines styles and colors for Zefyr editor toolbar.
-class ToolbarTheme {
-  /// The background color of the toolbar.
-  final Color color;
+/// Style theme applied to a block of rich text, including single-line
+/// paragraphs.
+class TextBlockTheme {
+  /// Base text style for a text block.
+  final TextStyle style;
 
-  /// Color of buttons in toggled state.
-  final Color toggleColor;
+  /// Vertical spacing around a text block.
+  final VerticalSpacing spacing;
 
-  /// Color of button icons.
-  final Color iconColor;
+  /// Vertical spacing for individual lines within a text block.
+  ///
+  final VerticalSpacing lineSpacing;
 
-  /// Color of button icons in disabled state.
-  final Color disabledIconColor;
+  /// Decoration of a text block.
+  ///
+  /// Decoration, if present, is painted in the content area, excluding
+  /// any [spacing].
+  final BoxDecoration? decoration;
 
-  /// Creates default theme for editor toolbar.
-  factory ToolbarTheme.fallback(BuildContext context) {
-    final theme = Theme.of(context);
-    return ToolbarTheme._(
-      color: theme.primaryColorBrightness == Brightness.light
-          ? Colors.grey.shade300
-          : Colors.grey.shade800,
-      toggleColor: theme.primaryColorBrightness == Brightness.light
-          ? Colors.grey.shade400
-          : Colors.grey.shade900,
-      iconColor: theme.primaryIconTheme.color,
-      disabledIconColor: theme.disabledColor,
-    );
-  }
+  TextBlockTheme({
+    required this.style,
+    required this.spacing,
+    this.lineSpacing = const VerticalSpacing.zero(),
+    this.decoration,
+  });
+}
 
-  ToolbarTheme._({
-    @required this.color,
-    @required this.toggleColor,
-    @required this.iconColor,
-    @required this.disabledIconColor,
+/// Theme data for inline code.
+class InlineCodeThemeData {
+  /// Base text style for an inline code.
+  final TextStyle style;
+
+  /// Style override for inline code in headings level 1.
+  final TextStyle? heading1;
+
+  /// Style override for inline code in headings level 2.
+  final TextStyle? heading2;
+
+  /// Style override for inline code in headings level 3.
+  final TextStyle? heading3;
+
+  /// Background color for inline code.
+  final Color? backgroundColor;
+
+  /// Radius used when paining the background.
+  final Radius? radius;
+
+  InlineCodeThemeData({
+    required this.style,
+    this.heading1,
+    this.heading2,
+    this.heading3,
+    this.backgroundColor,
+    this.radius,
   });
 
-  /// Creates a new [ToolbarTheme] where each property from this object has
-  /// been merged with the matching property from the `other` object.
-  ToolbarTheme copyWith({
-    Color color,
-    Color toggleColor,
-    Color iconColor,
-    Color disabledIconColor,
-  }) {
-    return ToolbarTheme._(
-      color: color ?? this.color,
-      toggleColor: toggleColor ?? this.toggleColor,
-      iconColor: iconColor ?? this.iconColor,
-      disabledIconColor: disabledIconColor ?? this.disabledIconColor,
-    );
-  }
-
-  /// Creates a new [ToolbarTheme] where each property from this object has
-  /// been merged with the matching property from the `other` object.
-  ToolbarTheme merge(ToolbarTheme other) {
-    if (other == null) return this;
-    return copyWith(
-      color: other.color ?? color,
-      toggleColor: other.toggleColor ?? toggleColor,
-      iconColor: other.iconColor ?? iconColor,
-      disabledIconColor: other.disabledIconColor ?? disabledIconColor,
-    );
+  /// Returns effective style to use for inline code for the specified
+  /// [lineStyle].
+  TextStyle styleFor(NotusStyle lineStyle) {
+    if (lineStyle.containsSame(NotusAttribute.h1)) {
+      return heading1 ?? style;
+    } else if (lineStyle.containsSame(NotusAttribute.h2)) {
+      return heading2 ?? style;
+    } else if (lineStyle.containsSame(NotusAttribute.h3)) {
+      return heading3 ?? style;
+    }
+    return style;
   }
 
   @override
-  bool operator ==(other) {
-    if (other.runtimeType != runtimeType) return false;
-    final ToolbarTheme otherTheme = other;
-    return (otherTheme.color == color) &&
-        (otherTheme.toggleColor == toggleColor) &&
-        (otherTheme.iconColor == iconColor) &&
-        (otherTheme.disabledIconColor == disabledIconColor);
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other is! InlineCodeThemeData) return false;
+    return other.style == style &&
+        other.heading1 == heading1 &&
+        other.heading2 == heading2 &&
+        other.heading3 == heading3 &&
+        other.backgroundColor == backgroundColor &&
+        other.radius == radius;
   }
 
   @override
   int get hashCode =>
-      hashValues(color, toggleColor, iconColor, disabledIconColor);
+      Object.hash(style, heading1, heading2, heading3, backgroundColor, radius);
 }
