@@ -1,9 +1,10 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-import 'package:test/test.dart';
-import 'package:quill_delta/quill_delta.dart';
+
 import 'package:notus/notus.dart';
+import 'package:quill_delta/quill_delta.dart';
+import 'package:test/test.dart';
 
 final ul = NotusAttribute.ul.toJson();
 final bold = NotusAttribute.bold.toJson();
@@ -40,6 +41,22 @@ void main() {
         ..retain(1, unsetUl);
       expect(actual, expected);
     });
+
+    test('preserves last newline character', () {
+      final doc = Delta()..insert('\n');
+      final actual = rule.apply(doc, 0, 1);
+      final expected = Delta();
+      expect(actual, expected);
+    });
+
+    test('preserves last newline character on multi character delete', () {
+      final doc = Delta()..insert('Document\nTitle\n');
+      final actual = rule.apply(doc, 8, 7);
+      final expected = Delta()
+        ..retain(8)
+        ..delete(6);
+      expect(actual, expected);
+    });
   });
 
   group('$CatchAllDeleteRule', () {
@@ -53,16 +70,31 @@ void main() {
         ..delete(5);
       expect(actual, expected);
     });
+
+    test('preserves last newline character', () {
+      final doc = Delta()..insert('\n');
+      final actual = rule.apply(doc, 0, 1);
+      final expected = Delta();
+      expect(actual, expected);
+    });
+
+    test('preserves last newline character on multi character delete', () {
+      final doc = Delta()..insert('Document\n');
+      final actual = rule.apply(doc, 3, 6);
+      final expected = Delta()
+        ..retain(3)
+        ..delete(5);
+      expect(actual, expected);
+    });
   });
 
   group('$EnsureEmbedLineRule', () {
     final rule = EnsureEmbedLineRule();
 
     test('ensures line-break before embed', () {
-      final hr = NotusAttribute.embed.horizontalRule;
       final doc = Delta()
         ..insert('Document\n')
-        ..insert(kZeroWidthSpace, hr.toJson())
+        ..insert(BlockEmbed.horizontalRule)
         ..insert('\n');
       final actual = rule.apply(doc, 8, 1);
       final expected = Delta()..retain(8);
@@ -70,10 +102,9 @@ void main() {
     });
 
     test('ensures line-break after embed', () {
-      final hr = NotusAttribute.embed.horizontalRule;
       final doc = Delta()
         ..insert('Document\n')
-        ..insert(kZeroWidthSpace, hr.toJson())
+        ..insert(BlockEmbed.horizontalRule)
         ..insert('\n');
       final actual = rule.apply(doc, 10, 1);
       final expected = Delta()..retain(11);
@@ -81,12 +112,11 @@ void main() {
     });
 
     test('still deletes everything between embeds', () {
-      final hr = NotusAttribute.embed.horizontalRule;
       final doc = Delta()
         ..insert('Document\n')
-        ..insert(kZeroWidthSpace, hr.toJson())
+        ..insert(BlockEmbed.horizontalRule)
         ..insert('\nSome text\n')
-        ..insert(kZeroWidthSpace, hr.toJson())
+        ..insert(BlockEmbed.horizontalRule)
         ..insert('\n');
       final actual = rule.apply(doc, 10, 11);
       final expected = Delta()
@@ -96,10 +126,9 @@ void main() {
     });
 
     test('allows deleting empty line after embed', () {
-      final hr = NotusAttribute.embed.horizontalRule;
       final doc = Delta()
         ..insert('Document\n')
-        ..insert(kZeroWidthSpace, hr.toJson())
+        ..insert(BlockEmbed.horizontalRule)
         ..insert('\n')
         ..insert('\n', NotusAttribute.block.bulletList.toJson())
         ..insert('Text')
@@ -112,12 +141,11 @@ void main() {
     });
 
     test('allows deleting empty line(s) before embed', () {
-      final hr = NotusAttribute.embed.horizontalRule;
       final doc = Delta()
         ..insert('Document\n')
         ..insert('\n')
         ..insert('\n')
-        ..insert(kZeroWidthSpace, hr.toJson())
+        ..insert(BlockEmbed.horizontalRule)
         ..insert('\n')
         ..insert('Text')
         ..insert('\n');
