@@ -1,9 +1,9 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-import 'package:test/test.dart';
-import 'package:quill_delta/quill_delta.dart';
 import 'package:notus/notus.dart';
+import 'package:quill_delta/quill_delta.dart';
+import 'package:test/test.dart';
 
 final ul = NotusAttribute.ul.toJson();
 final bold = NotusAttribute.bold.toJson();
@@ -49,6 +49,23 @@ void main() {
         ..insert('Three!\n');
       final actual = rule.apply(doc, 7, 0, NotusAttribute.ul);
       final expected = Delta()..retain(9)..retain(1, ul);
+      expect(actual, expected);
+    });
+
+    test('removing checklist style from a line also removes checked style', () {
+      final cl = NotusAttribute.cl.toJson();
+      final checkedCl = Map<String, dynamic>.from(cl)
+        ..addAll(NotusAttribute.checked.toJson());
+      final doc = Delta()
+        ..insert('Title\nOne')
+        ..insert('\n', checkedCl)
+        ..insert('Two')
+        ..insert('\n', cl)
+        ..insert('End!\n');
+      final actual = rule.apply(doc, 7, 0, NotusAttribute.cl.unset);
+      final noBlockNoChecked = NotusAttribute.block.unset.toJson()
+        ..addAll(NotusAttribute.checked.unset.toJson());
+      final expected = Delta()..retain(9)..retain(1, noBlockNoChecked);
       expect(actual, expected);
     });
   });
